mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "blitz-agent-mcp"
version = "0.1.0"
description = "Blitz Agent MCP Server - Advanced sports database analysis and AI-powered insights"
authors = [{ name = "Tejas Srinivasan", email = "tejas@blitzagent.com" }]
readme = "README.md"
license = "MIT"
keywords = ["mcp", "server", "sports", "database", "ai", "analysis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastmcp>=0.2.0",
    "httpx>=0.27.0",
    "sqlalchemy[postgresql-asyncpg]>=2.0.29",
    "click>=8.1.7",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.1",
    "azure-cosmos>=4.7.0",
    "azure-identity>=1.16.1",
    "azure-search-documents==11.6.0b3",
    "firecrawl-py>=0.0.15",
    "openai>=1.3.3",
    "aiohttp>=3.9.5",
    "beautifulsoup4>=4.12.3",
    "unidecode>=1.3.8",
    "lxml>=5.2.2",
    "sportsdataverse>=0.0.3",
    "pandas>=2.2.2",
    "torch>=2.3.1",
    "matplotlib>=3.8.0",
    "seaborn>=0.12.2",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "plotly>=5.17.0",
    "rank_bm25>=0.2.2"
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "blitz-agent-mcp[dev]"
]

[project.scripts]
blitz-agent-mcp = "blitz_agent_mcp.main:main"

[project.urls]
Homepage = "https://github.com/tejassrinivasan/BlitzLLM"
Repository = "https://github.com/tejassrinivasan/BlitzLLM"
Issues = "https://github.com/tejassrinivasan/BlitzLLM/issues"

[tool.hatch.build.targets.wheel]
packages = ["blitz_agent_mcp"]

[tool.hatch.build.targets.sdist]
include = [
    "/blitz_agent_mcp",
    "/README.md",
    "/requirements.txt"
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
target-version = "py38"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.rye]
managed = true
dev-dependencies = []

[tool.pyright]
reportMissingImports = true
reportUnusedImport = true
pythonVersion = "3.9"
pythonPlatform = "All"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 